DS-562 move popoverapi directive to popover directory
