+++ ngx-fudis/projects/ngx-fudis/src/lib/components/description-list/description-list-item/description-list-item-term/description-list-item-term.component.ts
@@ -5,7 +5,7 @@ import { DescriptionListItemComponent } from '../description-list-item.component
 import { DescriptionListComponent } from '../../description-list.component';
 import { FudisIdService } from '../../../../services/id/id.service';
 import { BehaviorSubject } from 'rxjs';
-import { PopoverApiDirective } from '../../../../directives/tooltip/popover-api.directive';
+import { PopoverApiDirective } from '../../../../directives/popover/popover-api.directive';
 
 @Component({
   selector: 'fudis-dt',
