+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/fieldset/examples/fieldset-example.component.ts
@@ -4,7 +4,7 @@ import { FudisRadioButtonOption } from '../../../../types/forms';
 import { NgxFudisModule } from '../../../../ngx-fudis.module';
 import { CommonModule } from '@angular/common';
 import { FudisValidators } from '../../../../utilities/form/validators';
-import { PopoverApiDirective } from '../../../../directives/tooltip/popover-api.directive';
+import { PopoverApiDirective } from '../../../../directives/popover/popover-api.directive';
 
 interface MyRadioButton {
   value: string;
